third_party/WebKit/LayoutTests/external/wpt/css/css-transitions/events-004-expected.txt

This is a testharness.js-based test.
FAIL repeating lists assert_approx_equals: expected 0.02 +/- 0.0005 but got 0.03
FAIL truncating lists assert_approx_equals: expected 0.02 +/- 0.0005 but got 0.03
Harness: the test ran to completion.
